buildins printf_float -> float, new assembler tests

+++ doc/test_mCc/buildins.c
@@ -3,7 +3,7 @@
 void   __attribute__((cdecl)) print(const char *msg);
 void   __attribute__((cdecl)) print_nl(void);
 void   __attribute__((cdecl)) print_int(long x);
-void   __attribute__((cdecl)) print_float(double x);
+void   __attribute__((cdecl)) print_float(float x);
 long   __attribute__((cdecl)) read_int(void);
 double __attribute__((cdecl)) read_float(void);
 
@@ -22,7 +22,7 @@ void print_int(long x)
 	printf("%ld", x);
 }
 
-void print_float(double x)
+void print_float(float x)
 {
 	printf("%f", x);
 }
